pyproject.toml
[project]
name = "nicookie-monster-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "greenlet>=3.3.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
